-Added 20 onto the internal colour ids in order to accommodate El&L (ninni\dye_depot\registry\DDDyes.java)
-Changed 5 colour names throughout the entire mod:
 -Maroon -> Scarlet
 -Amber -> Golden
 -Mint -> Turquoise
 -Navy -> Phthalo
 -Indigo -> Royal

Technically only the ids needed to be changed. However, in doing so, this makes Dye Depot dependent on El & L's dyes. Not ideal.

---

In its current state this mod is unstable. Shulker boxes and Beds crash the game when attempting to render them in the inventory. This also applies for advancement toasts and in the world. I have commented out the entries for these two item groups in the creative inventory as a temporary crash dodger, but will need to find an actual fix for this to make it playable. Unfortunately I do not have the know how to actually do this.

(ninni\dye_depot\registry\DDCreativeModeTabs.java)